updating code

+++ app/scripts/services/push.js
@@ -11,9 +11,9 @@ angular.module('webAppApp')
   .service('PushNotifications', function (ENV, $http) {
 
     var applicationServerPublicKey = 'BLPfteECnB-dwfVxR1lg8cY33qsgT68siqs5Rf1QSARQrj9Z9NVnfJbG4xjd8IeA28sJB8WzjL02FZ-TR3Ov8sM'
-
     var isSubscribed = false;
     var swRegistration = null;
+    var hasSW = ('serviceWorker' in navigator && 'PushManager' in window);
 
     function urlB64ToUint8Array (base64String) {
       var padding = '='.repeat((4 - base64String.length % 4) % 4);
@@ -31,7 +31,7 @@ angular.module('webAppApp')
     }
 
     function subscribeOnServer (userId, subscription) {
-      if (userId && subscription) {
+      if (hasSW && userId && subscription) {
         var prom = $http({
           method: 'POST',
           url: ENV.apiEndpoint + '/push',
@@ -51,7 +51,7 @@ angular.module('webAppApp')
     }
 
     function unsubscribeOnServer (userId) {
-      if (userId) {
+      if (hasSW && userId) {
         var prom = $http({
           method: 'DELETE',
           url: ENV.apiEndpoint + '/push/' + userId
@@ -67,41 +67,45 @@ angular.module('webAppApp')
     }
 
     function subscribeUser (userId) {
-      var applicationServerKey = urlB64ToUint8Array(applicationServerPublicKey);
-      swRegistration.pushManager.subscribe({
-        userVisibleOnly: true,
-        applicationServerKey: applicationServerKey
-      })
-      .then(function(subscription) {
-        subscribeOnServer(userId, subscription)
-        isSubscribed = true;
-      })
-      .catch(function(err) {
-        console.log('Failed to subscribe the user: ', err);
-      });
+      if (hasSW) {
+        var applicationServerKey = urlB64ToUint8Array(applicationServerPublicKey);
+        swRegistration.pushManager.subscribe({
+          userVisibleOnly: true,
+          applicationServerKey: applicationServerKey
+        })
+        .then(function(subscription) {
+          subscribeOnServer(userId, subscription)
+          isSubscribed = true;
+        })
+        .catch(function(err) {
+          console.log('Failed to subscribe the user: ', err);
+        });
+      }
     }
 
     function unsubscribeUser(userId) {
-      swRegistration.pushManager.getSubscription()
-      .then(function(subscription) {
-        if (subscription) {
-          return subscription.unsubscribe();
-        }
-      })
-      .catch(function(error) {
-        console.log('Error unsubscribing', error);
-      })
-      .then(function() {
-        unsubscribeOnServer(userId)
-        console.log('User is unsubscribed.');
-        isSubscribed = false;
-      });
+      if (hasSW) {
+        swRegistration.pushManager.getSubscription()
+        .then(function(subscription) {
+          if (subscription) {
+            return subscription.unsubscribe();
+          }
+        })
+        .catch(function(error) {
+          console.log('Error unsubscribing', error);
+        })
+        .then(function() {
+          unsubscribeOnServer(userId)
+          console.log('User is unsubscribed.');
+          isSubscribed = false;
+        });
+      }
     }
 
     function init () {
-      if ('serviceWorker' in navigator && 'PushManager' in window) {
+      if (hasSW) {
         console.log('Service Worker and Push is supported')
-        navigator.serviceWorker.register('/sw.js')
+        navigator.serviceWorker.register('sw.js')
         .then(function(swReg) {
           console.log('Service Worker is registered', swReg)
           swRegistration = swReg
